backend/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
django-oauth-toolkit>=2.3.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.2
psycopg2-binary>=2.9.9
celery>=5.3.0
redis>=5.0.0
channels>=4.0.0
django-cors-headers>=4.2.0
Pillow>=10.0.0 